dev(lexer): extract stream and token mod
